refactor: reorganize utility imports and replace filePath utilities with fileTree equivalents
