+++ src/test/tabId.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from '@jest/globals';
-import { getNormalizedNotebookDiffId } from '../utils/tabId';
+import { getCustomTabId, getNormalizedNotebookDiffId, getNotebookTabId } from '../utils/tabId';
 
 describe('tab IDs', () => {
   test('keeps notebook diff resources and type distinct', () => {
@@ -24,6 +24,18 @@ describe('tab IDs', () => {
         modified,
         notebookType: 'jupyter-notebook',
       }),
+    );
+  });
+
+  test('creates stable custom and notebook IDs from their full resources', () => {
+    const customUri = 'vscode-remote://ssh-remote%2Bworkspace/project/example.custom';
+    const notebookUri = 'vscode-remote://ssh-remote%2Bworkspace/project/example.ipynb';
+
+    expect(getCustomTabId(customUri, 'example.custom')).toBe(
+      JSON.stringify({ uri: customUri, viewType: 'example.custom' }),
+    );
+    expect(getNotebookTabId(notebookUri, 'jupyter-notebook')).toBe(
+      JSON.stringify({ uri: notebookUri, notebookType: 'jupyter-notebook' }),
     );
   });
 });
